Factored WCS overlay functions out of session and into subobject
