+++ carta/session.py
@@ -10,11 +10,12 @@ import base64
 import posixpath
 
 from .image import Image
-from .constants import CoordinateSystem, LabelType, BeamType, PaletteColor, Overlay, PanelMode, GridMode, ComplexComponent, NumberFormat, Polarization
+from .constants import PanelMode, GridMode, ComplexComponent, Polarization
 from .backend import Backend
 from .protocol import Protocol
-from .util import logger, Macro, split_action_path, CartaBadID, CartaBadSession, CartaBadUrl
-from .validation import validate, String, Number, Color, Constant, Boolean, NoneOr, OneOf, IterableOf, MapOf, Union
+from .util import Macro, split_action_path, CartaBadID, CartaBadSession, CartaBadUrl
+from .validation import validate, String, Number, Color, Constant, Boolean, NoneOr, IterableOf, MapOf, Union
+from .wcs_overlay import WCSOverlay
 
 
 class Session:
@@ -41,6 +42,8 @@ class Session:
         The browser object associated with this session. This is created automatically when a new session is created with :obj:`carta.session.Session.create` or :obj:`carta.session.Session.start_and_create`.
     backend : :obj:`carta.backend.Backend`
         The backend object associated with this session. This is created automatically when a new session is created with :obj:`carta.session.Session.start_and_create`.
+    overlay : :obj:`carta.wcs_overlay.WCSOverlay`
+        Sub-object with functions related to the WCS overlay.
 
     Attributes
     ----------
@@ -54,8 +57,8 @@ class Session:
         self._browser = browser
         self._backend = backend
 
-        # This is a local point of reference for paths, and may not be in sync with the frontend's starting directory
-        self._pwd = None
+        # Sub-objects grouping related functions
+        self.overlay = WCSOverlay(self)
 
     def __del__(self):
         """Delete this session object."""
@@ -577,363 +580,6 @@ class Session:
         self.call_action("preferenceStore.setPreference", "imagePanelRows", rows)
         self.call_action("preferenceStore.setPreference", "imagePanelColumns", columns)
         self.call_action("preferenceStore.setPreference", "imagePanelMode", grid_mode)
-
-    # CANVAS AND OVERLAY
-    @validate(Number(), Number())
-    def set_view_area(self, width, height):
-        """Set the dimensions of the view area.
-
-        Parameters
-        ----------
-        width : {0}
-            The new width, in pixels, divided by the device pixel ratio.
-        height : {1}
-            The new height, in pixels, divided by the device pixel ratio.
-        """
-        self.call_action("overlayStore.setViewDimension", width, height)
-
-    @validate(Constant(CoordinateSystem))
-    def set_coordinate_system(self, system=CoordinateSystem.AUTO):
-        """Set the coordinate system.
-
-        Parameters
-        ----------
-        system : {0}
-            The coordinate system.
-        """
-        self.call_action("overlayStore.global.setSystem", system)
-
-    def coordinate_system(self):
-        """Get the coordinate system.
-
-        Returns
-        ----------
-        :obj:`carta.constants.CoordinateSystem`
-            The coordinate system.
-        """
-        return CoordinateSystem(self.get_value("overlayStore.global.system"))
-
-    @validate(Constant(LabelType))
-    def set_label_type(self, label_type):
-        """Set the label type.
-
-        Parameters
-        ----------
-        label_type : {0}
-            The label type.
-        """
-        self.call_action("overlayStore.global.setLabelType", label_type)
-
-    @validate(NoneOr(String()), NoneOr(String()), NoneOr(String()))
-    def set_text(self, title=None, label_x=None, label_y=None):
-        """Set custom title and/or the axis label text.
-
-        Parameters
-        ----------
-        title : {0}
-            The title text.
-        label_x : {1}
-            The X-axis text.
-        label_y : {2}
-            The Y-axis text.
-        """
-        if title is not None:
-            self.call_action("overlayStore.title.setCustomTitleString", title)
-            self.call_action("overlayStore.title.setCustomText", True)
-        if label_x is not None:
-            self.call_action("overlayStore.labels.setCustomLabelX", label_x)
-        if label_y is not None:
-            self.call_action("overlayStore.labels.setCustomLabelX", label_y)
-        if label_x is not None or label_y is not None:
-            self.call_action("overlayStore.labels.setCustomText", True)
-
-    def clear_text(self):
-        """Clear all custom title and axis text."""
-        self.call_action("overlayStore.title.setCustomText", False)
-        self.call_action("overlayStore.labels.setCustomText", False)
-
-    @validate(OneOf(Overlay.TITLE, Overlay.NUMBERS, Overlay.LABELS), NoneOr(String()), NoneOr(Number()))
-    def set_font(self, component, font=None, font_size=None):
-        """Set the font and/or font size of an overlay component.
-
-        TODO: can we get the allowed font names from somewhere?
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        font : {1}
-            The font name.
-        font_size : {2}
-            The font size.
-        """
-        if font is not None:
-            self.call_action(f"overlayStore.{component}.setFont", font)
-        if font_size is not None:
-            self.call_action(f"overlayStore.{component}.setFontSize", font_size)
-
-    @validate(NoneOr(Constant(NumberFormat)), NoneOr(Constant(NumberFormat)))
-    def set_custom_number_format(self, x_format=None, y_format=None):
-        """Set a custom X and Y number format.
-
-        Parameters
-        ----------
-        x_format : {0}
-            The X format. If this is unset, the last custom X format to be set will be restored.
-        x_format : {1}
-            The Y format. If this is unset, the last custom Y format to be set will be restored.
-        """
-        if x_format is not None:
-            self.call_overlay_action(Overlay.NUMBERS, "setFormatX", x_format)
-        if y_format is not None:
-            self.call_overlay_action(Overlay.NUMBERS, "setFormatY", y_format)
-        self.call_overlay_action(Overlay.NUMBERS, "setCustomFormat", True)
-
-    def clear_custom_number_format(self):
-        """Disable the custom X and Y number format."""
-        self.call_overlay_action(Overlay.NUMBERS, "setCustomFormat", False)
-
-    def number_format(self):
-        """Return the current X and Y number formats, and whether they are a custom setting.
-
-        If the image has no WCS information, both the X and Y formats will be ``None``.
-
-        If a custom number format is not set, the format is derived from the coordinate system.
-
-        Returns
-        -------
-        tuple (a member of :obj:`carta.constants.NumberFormat` or ``None``, a member of :obj:`carta.constants.NumberFormat` or ``None``, boolean)
-            A tuple containing the X format, the Y format, and whether a custom format is set.
-        """
-        number_format_x = self.get_overlay_value(Overlay.NUMBERS, "formatTypeX")
-        number_format_y = self.get_overlay_value(Overlay.NUMBERS, "formatTypeY")
-        custom_format = self.get_overlay_value(Overlay.NUMBERS, "customFormat")
-        return NumberFormat(number_format_x), NumberFormat(number_format_y), custom_format
-
-    @validate(NoneOr(Constant(BeamType)), NoneOr(Number()), NoneOr(Number()), NoneOr(Number()))
-    def set_beam(self, beam_type=None, width=None, shift_x=None, shift_y=None):
-        """Set the beam properties.
-
-        Parameters
-        ----------
-        beam_type : {0}
-            The beam type.
-        width : {1}
-            The beam width.
-        shift_x : {2}
-            The X position.
-        shift_y : {3}
-            The Y position.
-        """
-        if beam_type is not None:
-            self.call_action(f"overlayStore.{Overlay.BEAM}.setBeamType", beam_type)
-        if width is not None:
-            self.call_action(f"overlayStore.{Overlay.BEAM}.setWidth", width)
-        if shift_x is not None:
-            self.call_action(f"overlayStore.{Overlay.BEAM}.setShiftX", shift_x)
-        if shift_y is not None:
-            self.call_action(f"overlayStore.{Overlay.BEAM}.setShiftY", shift_y)
-
-    @validate(Constant(PaletteColor), Constant(Overlay))
-    def set_color(self, color, component=Overlay.GLOBAL):
-        """Set the custom color on an overlay component, or the global color.
-
-        Parameters
-        ----------
-        color : {0}
-            The color.
-        component : {1}
-            The overlay component.
-        """
-        self.call_action(f"overlayStore.{component}.setColor", color)
-        if component not in (Overlay.GLOBAL, Overlay.BEAM):
-            self.call_action(f"overlayStore.{component}.setCustomColor", True)
-
-    @validate(Constant(Overlay, exclude=(Overlay.GLOBAL,)))
-    def clear_color(self, component):
-        """Clear the custom color from an overlay component.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        """
-        if component == Overlay.BEAM:
-            logger.warning("Cannot clear the color from the beam component. A color must be set on this component explicitly.")
-            return
-
-        self.call_action(f"overlayStore.{component}.setCustomColor", False)
-
-    @validate(Constant(Overlay))
-    def color(self, component):
-        """The color of an overlay component.
-
-        If called on the global overlay options, this function returns the global (default) overlay color. For any single overlay component other than the beam, it returns its custom color if a custom color is enabled, otherwise None. For the beam it returns the beam color.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-
-        Returns
-        -------
-        A member of :obj:`carta.constants.PaletteColor` or None
-            The color of the component or None if no custom color is set on the component.
-        """
-        if component in (Overlay.GLOBAL, Overlay.BEAM) or self.get_value(f"overlayStore.{component}.customColor"):
-            return PaletteColor(self.get_value(f"overlayStore.{component}.color"))
-
-    @validate(Constant(PaletteColor))
-    def palette_to_rgb(self, color):
-        """Convert a palette colour to RGB.
-
-        The RGB value depends on whether the session is using the light theme or the dark theme.
-
-        Parameters
-        ----------
-        color : {0}
-            The colour to convert.
-
-        Returns
-        -------
-        string
-            The RGB value of the palette colour in the session's current theme, as a 6-digit hexadecimal with a leading ``#``.
-        """
-        color = PaletteColor(color)
-        if self.get_value("darkTheme"):
-            return color.rgb_dark
-        return color.rgb_light
-
-    @validate(Number(min=0, interval=Number.EXCLUDE), OneOf(Overlay.GRID, Overlay.BORDER, Overlay.TICKS, Overlay.AXES, Overlay.COLORBAR))
-    def set_width(self, width, component):
-        """Set the line width of an overlay component.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        """
-        self.call_action(f"overlayStore.{component}.setWidth", width)
-
-    @validate(OneOf(Overlay.GRID, Overlay.BORDER, Overlay.TICKS, Overlay.AXES, Overlay.COLORBAR))
-    def width(self, component):
-        """The line width of an overlay component.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-
-        Returns
-        ----------
-        number
-            The line width of the component.
-        """
-        return self.get_value(f"overlayStore.{component}.width")
-
-    @validate(Constant(Overlay, exclude=(Overlay.GLOBAL,)), Boolean())
-    def set_visible(self, component, visible):
-        """Set the visibility of an overlay component.
-
-        Ticks cannot be shown or hidden in AST, but it is possible to set the width to a very small non-zero number to make them effectively invisible.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        visible : {1}
-            The visibility state.
-        """
-        if component == Overlay.TICKS:
-            logger.warning("Ticks cannot be shown or hidden.")
-            return
-
-        self.call_action(f"overlayStore.{component}.setVisible", visible)
-
-    @validate(Constant(Overlay, exclude=(Overlay.GLOBAL,)))
-    def visible(self, component):
-        """Whether an overlay component is visible.
-
-        Ticks cannot be shown or hidden in AST.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-
-        Returns
-        -------
-        boolean or None
-            Whether the component is visible, or None for an invalid component.
-        """
-        if component == Overlay.TICKS:
-            logger.warning("Ticks cannot be shown or hidden.")
-            return
-
-        return self.get_value(f"overlayStore.{component}.visible")
-
-    @validate(Constant(Overlay, exclude=(Overlay.GLOBAL,)))
-    def show(self, component):
-        """Show an overlay component.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        """
-        self.set_visible(component, True)
-
-    @validate(Constant(Overlay, exclude=(Overlay.GLOBAL,)))
-    def hide(self, component):
-        """Hide an overlay component.
-
-        Parameters
-        ----------
-        component : {0}
-            The overlay component.
-        """
-        self.set_visible(component, False)
-
-    def call_overlay_action(self, component, path, *args, **kwargs):
-        """Helper method for calling overlay component actions.
-
-        This method calls :obj:`carta.session.Session.call_action` after prepending this component's base path to the path parameter.
-
-        Parameters
-        ----------
-        component : a member of :obj:`carta.constants.Overlay`
-            The overlay component to use as the base of the path.
-        path : string
-            The path to an action relative to this overlay component.
-        *args
-            A variable-length list of parameters. These are passed unmodified to :obj:`carta.Session.call_action`.
-        **kwargs
-            Arbitrary keyword parameters. These are passed unmodified to :obj:`carta.Session.call_action`.
-        """
-        self.call_action(f"overlayStore.{component}.{path}", *args, **kwargs)
-
-    def get_overlay_value(self, component, path):
-        """Helper method for retrieving the values of overlay component attributes.
-
-        This method calls :obj:`carta.session.Session.get_value` after prepending this component's base path to the path parameter.
-
-        Parameters
-        ----------
-        component : a member of :obj:`carta.constants.Overlay`
-            The overlay component to use as the base of the path.
-        path : string
-            The path to an attribute relative to this overlay component.
-
-        Returns
-        -------
-        object
-            The value of the attribute, deserialized from a JSON string.
-        """
-        return self.get_value(f"overlayStore.{component}.{path}")
-
-    def toggle_labels(self):
-        """Toggle the overlay labels."""
-        self.call_action("overlayStore.toggleLabels")
 
     # PROFILES (TODO)
